lemonades: Don't process display brightness on pxlw disp mode change
In libpwirissoft.so's `pxlw` class, `qcomSetDisplayMode()` ends up
calling `irisSetConfigEx()` with `brightness_level_` as one of the
parameters. `irisSetConfigEx()` ends up passing that `brightness_level_`
param to `processDisplayBrightness()`, which seems to do some sort of
additional dimming of the display output depending on what the current
brightness level is.

On AOSP, it looks like `brightness_level_` doesn't get set properly.
After hwcomposer is initialized for the first time (or terminated and
restarted), if we turn the panel off and back on at least once, and then
try changing display modes, hwcomposer seems to apply a constant dimming
of the display output regardless of the backlight level.

Let's work around that by nop-ing out the `irisSetConfigEx()` call in
`qcomSetDisplayMode()`. That way, `processDisplayBrightness()` won't get
called, and we will not get that dimming.

Change-Id: Ied87b229d596bcc597b9ccb0a47368cbcb97f54b